requirements.txt
aiohttp==3.7.4.post0
apturl==0.5.2
argcomplete==2.0.0
asgiref==3.4.1
asn1crypto==0.24.0
asttokens==2.0.4
async-timeout==3.0.1
attrs==21.2.0
base58==2.1.0
bitarray==1.2.2
black==22.3.0
Brlapi==0.6.6
certifi==2022.12.7
chardet==4.0.0
charset-normalizer==2.0.12
click==8.0.4
colorama==0.3.7
command-not-found==0.3
cryptography==2.1.4
cupshelpers==1.0
cytoolz==0.11.0
dataclasses==0.8
dataclassy==0.10.4
defer==1.0.6
distlib==0.3.4
distro-info===0.18ubuntu0.18.04.1
Django==3.2.18
django-admin-honeypot==1.1.0
django-admin-thumbnails==0.2.6
django-cors-headers==3.10.1
django-crispy-forms==1.13.0
django-environ==0.9.0
django-filter==21.1
django-multiselectfield==0.1.12
django-session-timeout==0.1.0
djangorestframework==3.13.1
djangorestframework-simplejwt==4.4.0
eip712==0.1.0
eth-abi==2.1.1
eth-account==0.5.5
eth-brownie==1.16.4
eth-event==1.2.3
eth-hash==0.3.2
eth-keyfile==0.5.1
eth-keys==0.3.3
eth-rlp==0.2.1
eth-typing==2.2.2
eth-utils==1.10.0
execnet==1.9.0
filelock==3.4.1
hexbytes==0.2.2
httplib2==0.11.3
hypothesis==6.21.6
idna==3.4
idna-ssl==1.1.0
importlib-metadata==4.8.3
importlib-resources==5.4.0
inflection==0.5.0
iniconfig==1.1.1
ipfshttpclient==0.8.0a2
jsonschema==3.2.0
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
lazy-object-proxy==1.6.0
louis==3.5.0
lru-dict==1.1.7
macaroonbakery==1.1.3
Mako==1.0.7
Markdown==3.3.7
MarkupSafe==1.0
meld==3.18.0
multiaddr==0.0.9
multidict==5.1.0
mypy-extensions==0.4.3
mythx-models==1.9.1
netaddr==0.8.0
netifaces==0.10.4
oauth==1.0.1
olefile==0.45.1
packaging==21.3
parsimonious==0.8.1
pathspec==0.9.0
pexpect==4.2.1
Pillow==5.1.0
pipx==1.0.0
platformdirs==2.4.0
pluggy==1.0.0
prompt-toolkit==3.0.20
protobuf==3.18.0
psutil==5.8.0
psycopg2-binary==2.9.3
py==1.10.0
py-solc-ast==1.2.9
py-solc-x==1.1.0
pycairo==1.16.2
pycrypto==2.6.1
pycryptodome==3.10.1
pycups==1.9.73
Pygments==2.10.0
pygments-lexer-solidity==0.7.0
pygobject==3.26.1
PyJWT==2.4.0
pymacaroons==0.13.0
PyNaCl==1.1.2
pyparsing==3.0.9
pyRFC3339==1.0
pyrsistent==0.18.0
pytest==6.2.5
pytest-forked==1.3.0
pytest-xdist==1.34.0
python-apt==1.6.6
python-dateutil==2.8.1
python-debian==0.1.32
python-decouple==3.8
python-dotenv==0.16.0
pythx==1.6.1
pytz==2022.7.1
pyxdg==0.25
PyYAML==5.4.1
regex==2021.8.28
reportlab==3.4.0
requests==2.27.1
requests-unixsocket==0.1.5
rlp==1.2.0
SecretStorage==2.3.1
semantic-version==2.8.5
simplejson==3.13.2
six==1.16.0
sortedcontainers==2.4.0
sqlparse==0.4.3
ssh-import-id==5.7
system-service==0.3
systemd-python==234
toml==0.10.2
tomli==1.2.3
toolz==0.11.1
tqdm==4.62.3
typed-ast==1.5.4
typing-extensions==4.1.1
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.26.14
usb-creator==0.3.3
userpath==1.7.0
varint==1.0.2
virtualenv==20.14.1
vvm==0.1.0
vyper==0.2.16
wadllib==1.3.2
wcwidth==0.2.5
web3==5.23.1
websockets==9.1
wrapt==1.12.1
xkit==0.0.0
yarl==1.6.3
zipp==3.6.0
zope.interface==4.3.2
